addresses comments made by Jason 1) change filepath so it works for everyone 2) rather than creating a new profile, pass in a profile interactor and use the username to let the profile interactor retrieve the profile 3) refactor user calories: delete the user calories entities and do all the computation in meal plan interactor 4) pass in recipe name rather than recipe in the mealplan controller, since controller shouldn't have direct access to entity; Instead, pass in a recipe interactor to the mealplan controller as well, so recipe interactor can look up the recipe by recipe name
Potential issue: need recipe and profile interactor to be created so I can use them.

Note: I didn't store the numbers for calculating calories as constant, because they are not expected to change. For the issue of accessing entity in gateway, I asked on Piazza and the instructor say it's fine: "the controller/presenter and gateway are treated differently. The gateway directly loads and saves entities, whereas the controller/presenter have additional infrastructure in place to prevent them from directly interacting with the entities — the Input/Ouput Boundaries/Data."